(WIP) Rework outputs, PHIs, and register allocation.
Moved "ports" to function nodes, and replaced the Op port with a PHI
type. Phi is only used when merging operations from more than one
op, as can be the case with output from Match sub-nodes. Updated block
compiler to more closely build basic blocks from the node structures,
simplifying the implementation and more closely integrating it with
the register allocation portion of the code.